+++ test/Falco.Tests/ResponseTests.fs
@@ -5,7 +5,6 @@ open System.Text.Json
 open System.Text.Json.Serialization
 open Falco
 open Falco.Markup
-open FSharp.Control.Tasks.V2.ContextInsensitive
 open FsUnit.Xunit
 open Microsoft.Net.Http.Headers
 open NSubstitute
@@ -34,7 +33,7 @@ let ``Response.withHeader should set header`` () =
         do! ctx
             |> (Response.withHeaders [ HeaderNames.Server, serverName ] >> Response.ofEmpty)
 
-        ctx.Response.Headers.[HeaderNames.Server]
+        ctx.Response.Headers.[HeaderNames.Server][0]
         |> should equal serverName
     }
 
@@ -47,7 +46,7 @@ let ``Response.withContentType should set header`` () =
         do! ctx
             |> (Response.withContentType contentType>> Response.ofEmpty)
 
-        ctx.Response.Headers.[HeaderNames.ContentType]
+        ctx.Response.Headers.[HeaderNames.ContentType][0]
         |> should equal contentType
     }
 
@@ -84,8 +83,8 @@ let ``Response.ofBinary produces valid inline result from Byte[]`` () =
 
         let! body = getResponseBody ctx
         let contentLength = ctx.Response.ContentLength
-        let contentType = ctx.Response.Headers.[HeaderNames.ContentType]
-        let contentDisposition = ctx.Response.Headers.[HeaderNames.ContentDisposition]
+        let contentType = ctx.Response.Headers.[HeaderNames.ContentType][0]
+        let contentDisposition = ctx.Response.Headers.[HeaderNames.ContentDisposition][0]
 
         body               |> should equal expected
         contentType        |> should equal contentType
@@ -104,8 +103,8 @@ let ``Response.ofAttachment produces valid attachment result from Byte[]`` () =
 
         let! body = getResponseBody ctx
         let contentLength = ctx.Response.ContentLength
-        let contentType = ctx.Response.Headers.[HeaderNames.ContentType]
-        let contentDisposition = ctx.Response.Headers.[HeaderNames.ContentDisposition]
+        let contentType = ctx.Response.Headers.[HeaderNames.ContentType][0]
+        let contentDisposition = ctx.Response.Headers.[HeaderNames.ContentDisposition][0]
 
         body               |> should equal expected
         contentType        |> should equal contentType
@@ -124,7 +123,7 @@ let ``Response.ofPlainText produces text/plain result`` () =
 
         let! body = getResponseBody ctx
         let contentLength = ctx.Response.ContentLength
-        let contentType = ctx.Response.Headers.[HeaderNames.ContentType]
+        let contentType = ctx.Response.Headers.[HeaderNames.ContentType][0]
 
         body          |> should equal expected
         contentLength |> should equal (int64 (Encoding.UTF8.GetByteCount(expected)))
@@ -143,7 +142,7 @@ let ``Response.ofJson produces applicaiton/json result`` () =
 
         let! body = getResponseBody ctx
         let contentLength = ctx.Response.ContentLength
-        let contentType = ctx.Response.Headers.[HeaderNames.ContentType]
+        let contentType = ctx.Response.Headers.[HeaderNames.ContentType][0]
 
         body          |> should equal expected
         contentLength |> should equal (int64 (Encoding.UTF8.GetByteCount(expected)))
@@ -165,7 +164,7 @@ let ``Response.ofJsonOptions produces applicaiton/json result ignoring nulls`` (
 
         let! body = getResponseBody ctx
         let contentLength = ctx.Response.ContentLength
-        let contentType = ctx.Response.Headers.[HeaderNames.ContentType]
+        let contentType = ctx.Response.Headers.[HeaderNames.ContentType][0]
 
         body          |> should equal expected
         contentLength |> should equal (int64 (Encoding.UTF8.GetByteCount(expected)))
@@ -191,7 +190,7 @@ let ``Response.ofHtml produces text/html result`` () =
 
         let! body = getResponseBody ctx
         let contentLength = ctx.Response.ContentLength
-        let contentType = ctx.Response.Headers.[HeaderNames.ContentType]
+        let contentType = ctx.Response.Headers.[HeaderNames.ContentType][0]
 
         body          |> should equal expected
         contentLength |> should equal (int64 (Encoding.UTF8.GetByteCount(expected)))
@@ -210,7 +209,7 @@ let ``Response.ofHtmlString produces text/html result`` () =
 
         let! body = getResponseBody ctx
         let contentLength = ctx.Response.ContentLength
-        let contentType = ctx.Response.Headers.[HeaderNames.ContentType]
+        let contentType = ctx.Response.Headers.[HeaderNames.ContentType][0]
 
         body          |> should equal expected
         contentLength |> should equal (int64 (Encoding.UTF8.GetByteCount(expected)))